⚡ refactor : pisahkan generate report ke service baru
